Teslo Shop
On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	modified:   03-admin-todos/.env.template
	modified:   03-admin-todos/README.md
	modified:   03-admin-todos/next.config.mjs
	new file:   03-admin-todos/nextauth.d.ts
	modified:   03-admin-todos/package-lock.json
	modified:   03-admin-todos/package.json
	new file:   03-admin-todos/prisma/migrations/20241115202749_authjs/migration.sql
	new file:   03-admin-todos/prisma/migrations/20241115204116_nextauthjs/migration.sql
	new file:   03-admin-todos/prisma/migrations/20241115205155_nextauthjs/migration.sql
	new file:   03-admin-todos/prisma/migrations/20241121172055_roles_is_active/migration.sql
	new file:   03-admin-todos/prisma/migrations/20241121190442_password/migration.sql
	new file:   03-admin-todos/prisma/migrations/20241126171028_user_todos/migration.sql
	modified:   03-admin-todos/prisma/schema.prisma
	new file:   03-admin-todos/public/images/products/1623735-00-A_0_2000.jpg
	new file:   03-admin-todos/public/images/products/1657916-00-A_1.jpg
	new file:   03-admin-todos/public/images/products/1700280-00-A_1.jpg
	new file:   03-admin-todos/public/images/products/1742702-00-A_0_2000.jpg
	new file:   03-admin-todos/src/app/api/auth/[...nextauth]/route.ts
	modified:   03-admin-todos/src/app/api/seed/route.ts
	modified:   03-admin-todos/src/app/api/todos/[id]/route.ts
	modified:   03-admin-todos/src/app/api/todos/route.ts
	new file:   03-admin-todos/src/app/dashboard/cart/page.tsx
	new file:   03-admin-todos/src/app/dashboard/cookies/page.tsx
	modified:   03-admin-todos/src/app/dashboard/layout.tsx
	modified:   03-admin-todos/src/app/dashboard/page.tsx
	new file:   03-admin-todos/src/app/dashboard/products/page.tsx
	new file:   03-admin-todos/src/app/dashboard/profile/page.tsx
	modified:   03-admin-todos/src/app/dashboard/rest-todos/page.tsx
	modified:   03-admin-todos/src/app/dashboard/server-todos/page.tsx
	modified:   03-admin-todos/src/app/layout.tsx
	new file:   03-admin-todos/src/auth/actions/auth-actions.ts

+++ 03-admin-todos/src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServer } from "@/auth/actions/auth-actions";
 import prisma from "@/lib/prisma";
 import { NextResponse } from "next/server";
 import * as yup from "yup";
@@ -41,11 +42,18 @@ const postSchema = yup.object({
 
 
 export async function POST(request: Request) {
+    const user = await getUserSessionServer();
+    if (!user) {
+        return NextResponse.json('No autorizado', {status: 401});
+    }
+
+
     try {
         const { complete, description } = await postSchema.validate(await request.json());
         const todo = await prisma.todo.create({ data: { 
             complete, 
-            description
+            description,
+            userId: user.id
         }});
 
 
@@ -61,11 +69,18 @@ const deleteSchema = yup.object({
 });
 
 export async function DELETE(request: Request) {
+    const user = await getUserSessionServer();
+    if (!user) {
+        return NextResponse.json('No autorizado', {status: 401});
+    }
+
+
     try {
         const { complete } = await deleteSchema.validate(await request.json());
         const todosDelete = await prisma.todo.deleteMany({
             where: {
-                complete
+                complete,
+                userId: user.id
             }
         });
